In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,10 +6,10 @@
  */
 int main(void)
 {
-	unsigned int a = 1;
-	unsigned int b = 2;
+	long unsigned int a = 1;
+	long unsigned int b = 2;
 	int x;
-	unsigned int res;
+	long unsigned int res;
 	for (x = 1; x <= 98; x++)
 	{
 		printf("%lu, ", a);
